# openapi.yaml
openapi: 3.1.0
info:
  title: 国学老师 H5 API
  version: 2.7.0
  description: 首页与专业排盘公开接口。出生或起盘信息和排盘引用仅通过 POST 请求体提交。
paths:
...
  /api/v1/paipan/areas:
    get:
      summary: 获取排盘地区树
      responses:
        "200":
          description: 地区树，其他地区的编码为 999999
          headers:
            Cache-Control:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AreaNode"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/bazi/resolve-birth:
    post:
      summary: 将阳历、阴历或四柱反查解析为阳历候选
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveBirthRequest"
      responses:
        "200":
          description: 候选阳历时间；四柱反查固定使用 sect=2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveBirthResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/bazi/chart:
    post:
      summary: 生成生平子时完整命盘
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChartRequest"
      responses:
        "200":
          description: 分组后的完整排盘结果及短期排盘引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/bazi/context:
    post:
      summary: 使用短期排盘引用读取完整排盘上下文
      description: >-
        为页面刷新恢复、未来智能老师后端及 MCP 工具提供同一份强类型数据。
        paipan_ref 是短期读取凭证，不应记录到日志或拼接原始出生信息。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref:
                  $ref: "#/components/schemas/PaipanReference"
      responses:
        "200":
          description: 完整、带版本号的排盘上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaipanContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/bazi/flow-months:
    post:
      summary: 获取指定流年的十二流月
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chart, year]
              properties:
                chart:
                  $ref: "#/components/schemas/ChartRequest"
                year:
                  type: integer
                  minimum: 1900
                  maximum: 2200
      responses:
        "200":
          description: 十二流月及联动信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowMonthsResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/bazi/shen-sha:
    post:
      summary: 按盘面目标计算神煞
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pillars, target]
              properties:
                pillars:
                  type: object
                  additionalProperties:
                    type: array
                    minItems: 2
                    maxItems: 2
                    items:
                      type: string
                      minLength: 1
                      maxLength: 1
                target:
                  $ref: "#/components/schemas/ShenShaTarget"
      responses:
        "200":
          description: 神煞名称
          content:
            application/json:
              schema:
                type: object
                required: [target, names]
                properties:
                  target:
                    $ref: "#/components/schemas/ShenShaTarget"
                  names:
                    type: array
                    items: { type: string }
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/dunjia/chart:
    post:
      summary: 生成转盘拆补遁甲盘
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DunjiaChartRequest"
      responses:
        "200":
          description: 九宫、节气、旬首、四害、长生及短期排盘引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DunjiaChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/dunjia/context:
    post:
      summary: 使用短期排盘引用读取完整遁甲上下文
      description: >-
        为页面刷新恢复、未来智能老师后端及 MCP 工具提供同一份强类型数据。
        paipan_ref 是短期读取凭证，不应记录到日志或拼接原始起盘信息。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref:
                  $ref: "#/components/schemas/PaipanReference"
      responses:
        "200":
          description: 完整、带版本号的遁甲排盘上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DunjiaContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/juece/chart:
    post:
      summary: 生成时家决策盘
      description: 仅返回盘面结构，不提供个案吉凶判断。转盘为兼容参考站固定按拆补、时空、寄坤宫及原始钟表时间起局；飞盘按请求中的顺逆、定局、旬空和真太阳时条件起局。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JueceChartRequest"
      responses:
        "200":
          description: 标准化九宫盘及短期排盘引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JueceChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/juece/context:
    post:
      summary: 使用短期排盘引用读取时家决策上下文
      description: paipan_ref 只用于当前标签页短期恢复，服务端数据库仅保存引用哈希。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref:
                  $ref: "#/components/schemas/PaipanReference"
      responses:
        "200":
          description: 带盘型与版本号的完整时家决策上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JueceContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/yinpan-juece/chart:
    post:
      summary: 生成阴盘决策时盘或刻盘
      description: 保留参考站时盘、刻盘、事项、性别和可选终身局资料；结果只提供结构化盘面，不输出个案吉凶判断。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YinpanChartRequest"
      responses:
        "200":
          description: 阴盘九宫、四害、十二长生、天门地户及短期排盘引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YinpanChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/yinpan-juece/context:
    post:
      summary: 使用短期排盘引用读取阴盘决策上下文
      description: paipan_ref 只用于当前标签页短期恢复，服务端数据库仅保存引用哈希。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref:
                  $ref: "#/components/schemas/PaipanReference"
      responses:
        "200":
          description: 带盘型与版本号的完整阴盘决策上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YinpanContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/meihua/chart:
    post:
      summary: 生成梅花易数本卦、互卦与变卦
      description: 支持时间、随机、报数与指定四种起盘方式；随机起盘的上下卦及动爻由客户端生成后作为确定参数提交。六十四卦原文为前端静态参考资料，不进入排盘上下文。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeihuaChartRequest"
      responses:
        "200":
          description: 本卦、互卦、变卦、四柱信息及短期排盘引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeihuaChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/meihua/context:
    post:
      summary: 使用短期排盘引用读取梅花上下文
      description: paipan_ref 只用于当前标签页短期恢复，服务端数据库仅保存引用哈希。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref:
                  $ref: "#/components/schemas/PaipanReference"
      responses:
        "200":
          description: 带版本号的完整梅花排盘上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeihuaContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/luoji/chart:
    post:
      summary: 生成六爻逻辑盘
      description: 支持铜钱摇盘、盘名指定和六位硬币背数三种方式；返回本卦、变卦、八宫归属、世应、六神、伏神、六亲与纳甲，不输出个案吉凶判断。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LuojiChartRequest"
      responses:
        "200":
          description: 标准化六爻盘及短期排盘引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LuojiChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/luoji/context:
    post:
      summary: 使用短期排盘引用读取六爻逻辑盘上下文
      description: paipan_ref 只用于当前标签页短期恢复，服务端数据库仅保存引用哈希。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
      responses:
        "200":
          description: 带版本号的完整六爻逻辑盘上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LuojiContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/shanxiang-juece/chart:
    post:
      summary: 生成山向决策三盘
      description: 按输入角度所在十五度分区生成三个五度山向奇门盘，并返回短期恢复引用。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShanxiangChartRequest"
      responses:
        "200":
          description: 山向决策排盘结果与短期引用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShanxiangChartWithReference"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/paipan/shanxiang-juece/context:
    post:
      summary: 读取短期山向决策排盘上下文
      description: paipan_ref 只用于当前标签页短期恢复，服务端数据库仅保存引用哈希。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
      responses:
        "200":
          description: 带版本号的完整山向决策排盘上下文
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShanxiangContext"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/paipan/xingxiang/chart:
    post:
      summary: 生成飞星紫微盘
      description: 返回十二宫、星曜、生年四化、宫干自化、十二大限与每限十个流年，并签发短期排盘引用。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/XingxiangChartRequest" }
      responses:
        "200":
          description: 标准化飞星紫微盘及短期排盘引用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XingxiangChartWithReference" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "504": { $ref: "#/components/responses/GatewayTimeout" }
  /api/v1/paipan/xingxiang/context:
    post:
      summary: 使用短期排盘引用读取星像上下文
      description: paipan_ref 只用于当前标签页短期恢复，服务端数据库仅保存引用哈希。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
      responses:
        "200":
          description: 带版本号的完整星像排盘上下文
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XingxiangContext" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { $ref: "#/components/responses/Gone" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/paipan/shuzi-guilv/chart:
    post:
      summary: 生成先后天数字规律盘
      description: 按生肖、农历月日与时辰取数，返回先后天数字、阴阳五行及原系统特殊数组解读。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShuziGuilvChartRequest" }
      responses:
        "200":
          description: 标准化数字规律盘及短期排盘引用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShuziGuilvChartWithReference" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "504": { $ref: "#/components/responses/GatewayTimeout" }
  /api/v1/paipan/shuzi-guilv/context:
    post:
      summary: 使用短期排盘引用读取数字规律上下文
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
      responses:
        "200":
          description: 带版本号的完整数字规律排盘上下文
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShuziGuilvContext" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { $ref: "#/components/responses/Gone" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/paipan/xuankong-feixing/chart:
    post:
      summary: 生成玄空飞星九宫盘
      description: 按元运、二十四山、下盘或替盘规则布运星、山星、向星及年月日时飞星。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/XuankongFeixingChartRequest" }
      responses:
        "200":
          description: 标准化玄空飞星盘及短期排盘引用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XuankongFeixingChartWithReference" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "504": { $ref: "#/components/responses/GatewayTimeout" }
  /api/v1/paipan/xuankong-feixing/context:
    post:
      summary: 使用短期排盘引用读取玄空飞星上下文
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
      responses:
        "200":
          description: 带版本号的完整玄空飞星排盘上下文
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XuankongFeixingContext" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { $ref: "#/components/responses/Gone" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/paipan/xingming/chart:
    post:
      summary: 生成姓名学三才五格或六格盘
      description: 姓名按正式字典逐字规范化为康熙繁体，并使用严格康熙笔画计算；字库未收录的输入返回 422，正式数理资源缺项返回 500。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/XingmingChartRequest" }
      responses:
        "200":
          description: 含数据集状态的姓名学结果及短期排盘引用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XingmingChartWithReference" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "504": { $ref: "#/components/responses/GatewayTimeout" }
  /api/v1/paipan/xingming/context:
    post:
      summary: 使用短期排盘引用读取姓名学上下文
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paipan_ref]
              properties:
                paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
      responses:
        "200":
          description: 带版本号的完整姓名学上下文
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XingmingContext" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { $ref: "#/components/responses/Gone" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  responses:
    ValidationError:
      description: 字段校验或业务输入错误
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    BadGateway:
      description: 排盘算法服务不可用或返回异常
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    GatewayTimeout:
      description: 排盘算法服务超时
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: 排盘引用不存在
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Gone:
      description: 排盘引用已过期
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ShenShaTarget:
      type: string
      enum: [nian, yue, ri, shi, dayun, liunian, liuyue]
    AreaNode:
      type: object
      required: [label, code, children]
      properties:
        label: { type: string }
        code: { type: string }
        children:
          type: array
          items:
            $ref: "#/components/schemas/AreaNode"
    ResolveBirthRequest:
      oneOf:
        - type: object
          required: [mode, solarDateTime]
          properties:
            mode: { const: solar }
            solarDateTime:
              $ref: "#/components/schemas/BirthDateTime"
        - type: object
          required: [mode, lunar]
          properties:
            mode: { const: lunar }
            lunar:
              type: object
              required: [year, month, day, hour, minute, leapMonth]
              properties:
                year: { type: integer, minimum: 1900, maximum: 2100 }
                month: { type: integer, minimum: 1, maximum: 12 }
                day: { type: integer, minimum: 1, maximum: 30 }
                hour: { type: integer, minimum: 0, maximum: 23 }
                minute: { type: integer, minimum: 0, maximum: 59 }
                leapMonth: { type: boolean }
        - type: object
          required: [mode, pillars]
          properties:
            mode: { const: fourPillars }
            pillars:
              type: object
              required: [year, month, day, hour]
              properties:
                year: { type: string, minLength: 2, maxLength: 2 }
                month: { type: string, minLength: 2, maxLength: 2 }
                day: { type: string, minLength: 2, maxLength: 2 }
                hour: { type: string, minLength: 2, maxLength: 2 }
      discriminator:
        propertyName: mode
    ResolveBirthResponse:
      type: object
      required: [candidates, sect]
      properties:
        candidates:
          type: array
          items:
            type: object
            required: [id, solarDateTime, label]
            properties:
              id: { type: string }
              solarDateTime:
                $ref: "#/components/schemas/BirthDateTime"
              label: { type: string }
        sect: { const: 2 }
    ChartRequest:
      type: object
      required: [name, gender, birthDateTime, areaCode, useTrueSolarTime]
      properties:
        name: { type: string, maxLength: 32, description: 可为空字符串 }
        gender: { type: string, enum: [male, female] }
        birthDateTime:
          $ref: "#/components/schemas/BirthDateTime"
        areaCode: { type: string, pattern: '^\d{6}$' }
        useTrueSolarTime: { type: boolean }
...
    PaipanReference:
      type: string
      pattern: '^pp_[A-Za-z0-9_-]{32}$'
      example: pp_0123456789abcdefghijklmnopqrstuv
      description: 具有读取能力的短期随机引用；只通过 POST 请求体或受控 Dify/MCP 输入传递
    ChartWithReference:
      allOf:
        - $ref: "#/components/schemas/ChartResponse"
        - type: object
          required: [paipan_ref, expiresAt]
          properties:
            paipan_ref:
              $ref: "#/components/schemas/PaipanReference"
            expiresAt:
              type: string
              format: date-time
    PaipanContext:
      type: object
      required: [schemaVersion, chartType, paipan_ref, generatedAt, expiresAt, chartRequest, chart]
      properties:
        schemaVersion:
          const: guoxue.paipan.bazi.v1
        chartType:
          const: shengping_zishi
        paipan_ref:
          $ref: "#/components/schemas/PaipanReference"
        generatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        chartRequest:
          $ref: "#/components/schemas/ChartRequest"
        chart:
          $ref: "#/components/schemas/ChartResponse"
    DunjiaChartRequest:
      type: object
      required: [chartDateTime]
      properties:
        chartDateTime:
          $ref: "#/components/schemas/BirthDateTime"
...
    DunjiaChartWithReference:
      allOf:
        - $ref: "#/components/schemas/DunjiaChartResponse"
        - type: object
          required: [paipan_ref, expiresAt]
          properties:
            paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
            expiresAt: { type: string, format: date-time }
    DunjiaContext:
      type: object
      required: [schemaVersion, chartType, paipan_ref, generatedAt, expiresAt, chartRequest, chart]
      properties:
        schemaVersion: { const: guoxue.paipan.dunjia.v1 }
        chartType: { const: dunjia }
        paipan_ref: { $ref: "#/components/schemas/PaipanReference" }
        generatedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        chartRequest: { $ref: "#/components/schemas/DunjiaChartRequest" }
        chart: { $ref: "#/components/schemas/DunjiaChartResponse" }
...
    ShanxiangChartRequest:
      type: object
      additionalProperties: false
      required: [year, degrees]
      properties:
        year: { type: integer, minimum: 1930, maximum: 2100 }
        degrees: { type: number, minimum: 0, maximum: 360 }
        question: { type: string, maxLength: 80, default: "" }
...
    FlowMonthsResponse:
      type: object
      required: [year, months]
      properties:
        year: { type: integer }
        months:
          type: array
          minItems: 12
          maxItems: 12
          items: { $ref: "#/components/schemas/FlowMonth" }
...